Archivos-Configuracion-Practica-2/requirements.txt
python-decouple==3.6
python-dotenv==0.21.0
